requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
requests==2.31.0
slack-sdk==3.35.0
groq==0.19.0